Found the problem
Registers ascend at the end of the raw function execution, but
they also need to be ascended at the end of the function call
(very bottom of call.go). But when I put it there, it hung
because I wasn't storing a return location anywhere--I wasn't
remember-jumping, and I was just looping forever.

But I realized while doing this that the pre-executed nature of my
debugger means that I can't debug code that loops forever. So I
updated it to interpret as it goes. test.go is full of dead code now;
clear it out soon.